pyproject.toml
[tool.poetry]
name = "postgres_pubsub"
version = "0.1.0"
description = ""
authors = ["Ville Kärkkäinen <ville.karkkainen@outlook.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
psycopg2-binary = "2.8.6"

[tool.poetry.dev-dependencies]
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
